backend/requirements.txt
Flask
python-dotenv     
Flask-Cors         
Flask-RESTful    
requests
openai
flask
flask-cors